recursively replace embedded tokens
